# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/projects/review-queue": {
      "get": {
        "operationId": "AdminProjectsController_reviewQueue_v1",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": { "type": "string", "enum": ["DRAFT", "IN_REVIEW", "PUBLISHED", "ARCHIVED"] }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": { "maxLength": 100, "type": "string" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque keyset cursor",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminProjectListResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Admin Projects"]
      }
    },
    "/api/v1/admin/projects/{id}/status": {
      "patch": {
        "operationId": "AdminProjectsController_updateStatus_v1",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateAdminProjectStatusDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminProjectMutationResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Admin Projects"]
      }
    },
    "/api/v1/admin/overview": {
      "get": {
        "operationId": "AdminOperationsController_overview_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminOverviewResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Admin Project Operations"]
      }
    },
    "/api/v1/admin/data-quality": {
      "get": {
        "operationId": "AdminOperationsController_dataQuality_v1",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": { "type": "string", "enum": ["DRAFT", "IN_REVIEW", "PUBLISHED", "ARCHIVED"] }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": { "maxLength": 100, "type": "string" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque keyset cursor",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminProjectListResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Admin Project Operations"]
      }
    },
...
  },
  "info": {
    "title": "PLANDA Product API",
    "description": "Shared API for web, iOS, Android, developer, broker and admin clients",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 400 },
          "code": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string", "example": "Validation failed" },
          "details": { "type": "array", "items": { "type": "string" } },
          "requestId": { "type": "string", "example": "4f12bc5f-1a5e-4d14-bf14-6c7bb19ea413" }
        },
        "required": ["statusCode", "code", "message", "requestId"]
      },
...
      "Object": { "type": "object", "properties": {} },
      "OperationsPageInfoDto": {
        "type": "object",
        "properties": {
          "nextCursor": { "type": "object", "nullable": true },
          "hasNextPage": { "type": "boolean" }
        },
        "required": ["hasNextPage"]
      },
...
      "ProjectLocationReferenceDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" }
        },
        "required": ["id", "code", "name", "slug"]
      },
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "developerOrganizationId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "developerName": { "type": "string" },
          "developerVerified": { "type": "boolean" },
          "status": { "type": "string", "enum": ["DRAFT", "IN_REVIEW", "PUBLISHED", "ARCHIVED"] },
          "constructionStatus": {
            "type": "string",
            "enum": ["PLANNED", "UNDER_CONSTRUCTION", "READY"]
          },
          "province": { "$ref": "#/components/schemas/ProjectLocationReferenceDto" },
          "district": { "$ref": "#/components/schemas/ProjectLocationReferenceDto" },
          "latitude": { "type": "string", "example": "40.990868" },
          "longitude": { "type": "string", "example": "29.027707" },
          "startingPrice": { "type": "string", "example": "7500000.0000" },
          "currency": { "type": "string", "example": "TRY" },
          "deliveryDate": { "type": "string", "format": "date", "nullable": true },
          "summary": { "type": "string", "nullable": true },
          "heroImageUrl": { "type": "string", "nullable": true },
          "stockUpdatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "priceUpdatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "version": { "type": "number" }
        },
        "required": [
          "id",
          "developerOrganizationId",
          "name",
          "slug",
          "developerName",
          "developerVerified",
          "status",
          "constructionStatus",
          "province",
          "district",
          "latitude",
          "longitude",
          "startingPrice",
          "currency",
          "deliveryDate",
          "summary",
          "heroImageUrl",
          "stockUpdatedAt",
          "priceUpdatedAt",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
...
      "AdminProjectReadItemDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "developerOrganizationId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "developerName": { "type": "string" },
          "developerVerified": { "type": "boolean" },
          "status": { "type": "string", "enum": ["DRAFT", "IN_REVIEW", "PUBLISHED", "ARCHIVED"] },
          "constructionStatus": {
            "type": "string",
            "enum": ["PLANNED", "UNDER_CONSTRUCTION", "READY"]
          },
          "province": { "$ref": "#/components/schemas/ProjectLocationReferenceDto" },
          "district": { "$ref": "#/components/schemas/ProjectLocationReferenceDto" },
          "latitude": { "type": "string", "example": "40.990868" },
          "longitude": { "type": "string", "example": "29.027707" },
          "startingPrice": { "type": "string", "example": "7500000.0000" },
          "currency": { "type": "string", "example": "TRY" },
          "deliveryDate": { "type": "string", "format": "date", "nullable": true },
          "summary": { "type": "string", "nullable": true },
          "heroImageUrl": { "type": "string", "nullable": true },
          "stockUpdatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "priceUpdatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "version": { "type": "number" },
          "unitCount": { "type": "number" },
          "availableUnitCount": { "type": "number" },
          "leadCount": { "type": "number" },
          "paymentPlanCount": { "type": "number" },
          "completenessPercent": { "type": "number", "minimum": 0, "maximum": 100 },
          "issues": { "type": "array", "items": { "type": "string" } }
        },
        "required": [
          "id",
          "developerOrganizationId",
          "name",
          "slug",
          "developerName",
          "developerVerified",
          "status",
          "constructionStatus",
          "province",
          "district",
          "latitude",
          "longitude",
          "startingPrice",
          "currency",
          "deliveryDate",
          "summary",
          "heroImageUrl",
          "stockUpdatedAt",
          "priceUpdatedAt",
          "createdAt",
          "updatedAt",
          "version",
          "unitCount",
          "availableUnitCount",
          "leadCount",
          "paymentPlanCount",
          "completenessPercent",
          "issues"
        ]
      },
      "AdminProjectListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AdminProjectReadItemDto" }
          },
          "pageInfo": { "$ref": "#/components/schemas/OperationsPageInfoDto" }
        },
        "required": ["items", "pageInfo"]
      },
      "UpdateAdminProjectStatusDto": {
        "type": "object",
        "properties": { "status": { "type": "string", "enum": ["PUBLISHED", "ARCHIVED"] } },
        "required": ["status"]
      },
      "AdminProjectMutationResponseDto": {
        "type": "object",
        "properties": {
          "project": { "$ref": "#/components/schemas/ProjectResponseDto" },
          "auditId": { "type": "string", "format": "uuid" }
        },
        "required": ["project", "auditId"]
      },
      "AdminOverviewResponseDto": {
        "type": "object",
        "properties": {
          "publishedProjectCount": { "type": "number" },
          "inReviewProjectCount": { "type": "number" },
          "draftProjectCount": { "type": "number" },
          "archivedProjectCount": { "type": "number" },
          "staleStockProjectCount": { "type": "number" },
          "stalePriceProjectCount": { "type": "number" }
        },
        "required": [
          "publishedProjectCount",
          "inReviewProjectCount",
          "draftProjectCount",
          "archivedProjectCount",
          "staleStockProjectCount",
          "stalePriceProjectCount"
        ]
      },
...
    }
  }
}